added UserMapper, request class, refactored (#41)

+++ calendar/src/main/java/com/calendar/calendar/services/UserService.java
@@ -1,6 +1,8 @@
 package com.calendar.calendar.services;
 
+import com.calendar.calendar.dto.UserDTO;
 import com.calendar.calendar.entities.User;
+import com.calendar.calendar.mapper.UserMapper;
 import com.calendar.calendar.repositories.IUserRepository;
 import org.springframework.stereotype.Service;
 
@@ -10,20 +12,29 @@ import java.util.UUID;
 public class UserService {
 
     private final IUserRepository userRepository;
+    private final UserMapper userMapper;
 
-    public UserService(IUserRepository userRepository) {
+    public UserService(IUserRepository userRepository, UserMapper userMapper) {
         this.userRepository = userRepository;
+        this.userMapper = userMapper;
     }
 
-    public Boolean create(User user) {
-        User dbUser = userRepository.findByEmail(user.getEmail());
-        if (dbUser == null)
-            return userRepository.save(user) != null;
+    public Boolean create(UserDTO userDTO) {
+        User user = userMapper.toUser(userDTO);
+
+        if (userRepository.findByEmail(user.getEmail()) == null) {
+            userRepository.save(user);
+            return true;
+        }
         return false;
     }
 
-    public void delete(User user) {
+    public boolean delete(String email) {
+        User user = userRepository.findByEmail(email);
+        if (user == null)
+            return false;
         userRepository.delete(user);
+        return true;
     }
 
     public void deleteById(UUID id) {
@@ -34,7 +45,10 @@ public class UserService {
         userRepository.save(user);
     }
 
-    public User findUser(String email){
-        return userRepository.findByEmail(email);
+    public UserDTO findUser(String email) {
+        User user = userRepository.findByEmail(email);
+        if (user == null)
+            return null;
+        return userMapper.toUserDto(user);
     }
 }
